feat: help command listener feat: get ignored chats command feat: delete ignored chat command feat: made cache for db helpers feat: added message listener feat: made bot register all chats when joined refactor: made db helpers singletons

+++ src/main/java/db/IgnoredChatsDbHelper.java
@@ -3,12 +3,47 @@ package db;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.Statement;
+import java.util.*;
+import java.util.stream.Collectors;
 
-// TODO
+// TODO: check notion
 public class IgnoredChatsDbHelper {
+    private static IgnoredChatsDbHelper instance;
+    private final Map<String, List<String>> cache = new HashMap<>();
+
+    private IgnoredChatsDbHelper() {
+        List<Map.Entry<String, String>> allRows = getAllRows();
+
+        List<String> ids = allRows
+                .stream()
+                .map(Map.Entry::getKey)
+                .collect(Collectors.toList());
+
+        ids.forEach(serverId -> {
+                cache.put(serverId, new ArrayList<>());
+
+            List<String> chatIds = allRows.stream()
+                    .filter(pair -> pair.getKey().equals(serverId))
+                    .map(Map.Entry::getValue)
+                    .collect(Collectors.toList());
+
+            cache.get(serverId).addAll(chatIds);
+        });
+    }
+
+    public static synchronized IgnoredChatsDbHelper getInstance() {
+        if (instance == null) instance = new IgnoredChatsDbHelper();
+
+        return instance;
+    }
+
     public void insert(String serverId, String chatId) {
-        String sql = "insert into ignored_chats (server_id, chat_id) values (?, ?)";
+        cache.putIfAbsent(serverId, new ArrayList<>());
+        List<String> chats = cache.get(serverId);
+        if (!chats.contains(chatId)) chats.add(chatId);
+        else return;
+
+        String sql = "insert into ignored_chats (server_id, chat_id) values (?, ?) on conflict do nothing";
         Connection connection = DbUtil.getDbConnection();
         try (PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.setString(1, serverId);
@@ -21,23 +56,17 @@ public class IgnoredChatsDbHelper {
     }
 
     public boolean exists(String serverId, String chatId) {
-        boolean result = false;
-        String sql = "select case when count(*) > 0 then true else false end where server_id = ? and chat_id = ?";
-        Connection connection = DbUtil.getDbConnection();
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, serverId);
-            statement.setString(2, chatId);
-            ResultSet resultSet = statement.executeQuery();
-            result = resultSet.getBoolean(1);
-            connection.close();
-        } catch (Exception ex) {
-            System.err.println(ex.getMessage());
-        }
+        if (cache.get(serverId) == null) return false;
 
-        return result;
+        return cache.get(serverId).contains(chatId);
     }
 
     public void delete(String serverId, String chatId) {
+        if (cache.get(serverId) == null) return;
+        if (!cache.get(serverId).contains(chatId)) return;
+
+        cache.get(serverId).remove(chatId);
+
         String sql = "delete from ignored_chats where server_id = ? and chat_id = ?";
         Connection connection = DbUtil.getDbConnection();
         try (PreparedStatement statement = connection.prepareStatement(sql)) {
@@ -48,5 +77,34 @@ public class IgnoredChatsDbHelper {
         } catch (Exception ex) {
             System.err.println(ex.getMessage());
         }
+    }
+
+    public List<String> findIgnoredChatsByServerId(String serverId) {
+        // optimize later
+        return cache.get(serverId);
+    }
+
+    public void registerServer(String serverId) {
+        cache.putIfAbsent(serverId, new ArrayList<>());
+    }
+
+    private List<Map.Entry<String, String>> getAllRows() {
+        List<Map.Entry<String, String>> rows = new ArrayList<>();
+        String sql = "select server_id, chat_id from ignored_chats";
+        Connection connection = DbUtil.getDbConnection();
+        try(PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
+            ResultSet resultSet = preparedStatement.executeQuery();
+            while (resultSet.next()) {
+                String serverId = resultSet.getString(1);
+                String chatId = resultSet.getString(2);
+                rows.add(new AbstractMap.SimpleEntry<>(serverId, chatId));
+            }
+
+            connection.close();
+        } catch (Exception ex) {
+            System.out.println(ex.getMessage());
+        }
+
+        return rows;
     }
 }
